Fix tests

+++ SlangNet.ComWrappers/Slang.cs
@@ -9,14 +9,13 @@ namespace SlangNet.ComWrappers;
 public static partial class Slang
 {
     [LibraryImport("slang", EntryPoint = "slang_createGlobalSession")]
-    public static partial SlangResult CreateGlobalSession(nint apiVersion, out IGlobalSession globalSession);
+    public static partial SlangResult CreateGlobalSession(int apiVersion, out IGlobalSession globalSession);
 
     [LibraryImport("slang", EntryPoint = "slang_createGlobalSession2")]
     public static partial SlangResult CreateGlobalSession(GlobalSessionDescription desc, out IGlobalSession globalSession);
 
     [LibraryImport("slang", EntryPoint = "slang_createGlobalSessionWithoutCoreModule")]
-    public static partial SlangResult
-        CreateGlobalSessionWithoutCoreModule(nint apiVersion, out IGlobalSession globalSession);
+    public static partial SlangResult CreateGlobalSessionWithoutCoreModule(int apiVersion, out IGlobalSession globalSession);
 
     [LibraryImport("slang", EntryPoint = "slang_getEmbeddedCoreModule")]
     public static partial IBlob GetEmbeddedCoreModule();
